debugging
process the question that when the detectors is overlapping with each other (e.g., test the effect of the same SDS and different detector size), the replay function will assign the weight detected by the first probe to all detectable probe, regardless of the size.
The question is the for loop in the detect function.